+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 Improvements:
 
 - New feature: Permanent Buffs (did I merge that?)
+- New purchasable: More seed slots
+- Be able to obtain seeds from a single homozygous plant
 - Add some useful infos to the "a new bloom" message
 - Migrate Gray-violet to a tad more violet-ish tone
 - Expand orders book if there are new orders available